requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
Authlib==1.6.6
bcrypt==5.0.0
beartype==0.22.9
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cmdstanpy==1.3.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.4.0
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.126.0
fastmcp==2.14.1
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
h2==4.3.0
holidays==0.87
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.4.0
jiter==0.12.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
lupa==2.6
markdown-it-py==4.0.0
matplotlib==3.10.8
mcp==1.25.0
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
multidict==6.7.0
numpy==2.4.0
openai==2.14.0
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
passlib==1.7.2
pathable==0.4.4
pathvalidate==3.3.1
pillow==12.0.0
platformdirs==4.5.1
postgrest==2.27.0
prometheus_client==0.23.1
propcache==0.4.1
prophet==1.2.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.0
redis==7.1.0
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.0.4
stanio==0.5.1
starlette==0.50.0
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
tenacity==9.1.2
tqdm==4.67.1
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
azure-storage-blob==12.19.0
pandas>=2.0.0
fastapi>=0.104.0
python-multipart
openpyxl
